Moved about-child under the claim (rather than new-claim) directory.
